+++ eCommerce.Storefront.Model/Orders/IOrderRepository.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Orders
 {
-    public interface IOrderRepository : IRepository<Order, int>
+    public interface IOrderRepository : IRepository<Order, long>
     {         
     }
 }
