+++ eCommerce.Storefront.Model/Products/Brand.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Products
 {
-    public class Brand : EntityBase<int>, IAggregateRoot, IProductAttribute
+    public class Brand : EntityBase<long>, IAggregateRoot, IProductAttribute
     {
         public string Name { get; set; }
