+++ eCommerce.Storefront.Model/Products/IProductRepository.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Products
 {
-    public interface IProductRepository : IReadOnlyRepository<Product, int>
+    public interface IProductRepository : IReadOnlyRepository<Product, long>
     {
     }
 }
