+++ eCommerce.Storefront.Model/Products/IProductTitleRepository.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Products
 {
-    public interface IProductTitleRepository : IReadOnlyRepository<ProductTitle, int>
+    public interface IProductTitleRepository : IReadOnlyRepository<ProductTitle, long>
     {
     }
 }
