+++ eCommerce.Storefront.Model/Products/ProductColor.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Products
 {
-    public class ProductColor : EntityBase<int>, IAggregateRoot, IProductAttribute
+    public class ProductColor : EntityBase<long>, IAggregateRoot, IProductAttribute
     {
         public string Name { get; set; }
