+++ eCommerce.Storefront.Model/Products/ProductSize.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Products
 {
-    public class ProductSize : EntityBase<int>, IAggregateRoot, IProductAttribute
+    public class ProductSize : EntityBase<long>, IAggregateRoot, IProductAttribute
     {
         public string Name { get; set; }
