+++ eCommerce.Storefront.Model/Shipping/IDeliveryOptionRepository.cs
@@ -2,7 +2,7 @@ using Infrastructure.Domain;
 
 namespace eCommerce.Storefront.Model.Shipping
 {
-    public interface IDeliveryOptionRepository : IReadOnlyRepository<DeliveryOption, int>
+    public interface IDeliveryOptionRepository : IReadOnlyRepository<DeliveryOption, long>
     {
     }
 }
