+++ eCommerce.Storefront.Services/Messaging/ProductCatalogService/ModifyBasketRequest.cs
@@ -7,15 +7,15 @@ namespace eCommerce.Storefront.Services.Messaging.ProductCatalogService
     {
         public ModifyBasketRequest()
         {
-            ItemsToRemove = new List<int>();
-            ProductsToAdd = new List<int>();
+            ItemsToRemove = new List<long>();
+            ProductsToAdd = new List<long>();
             ItemsToUpdate = new List<ProductQtyUpdateRequest>();
         }
         
         public Guid BasketId { get; set; }
-        public IList<int> ItemsToRemove { get; set; }
+        public IList<long> ItemsToRemove { get; set; }
         public IList<ProductQtyUpdateRequest> ItemsToUpdate { get; set; }
         public int SetShippingServiceIdTo { get; set; }
-        public IList<int> ProductsToAdd { get; set; }
+        public IList<long> ProductsToAdd { get; set; }
     }
 }
